requirements.txt
requests
beautifulsoup4
sentence-transformers
pymilvus
transformers
torch
